PresentationWorkFlow/requirements.txt
# Python依存関係

# YAML処理
PyYAML>=6.0.1

# Google Generative AI API (Gemini)
google-genai>=0.5.0

# 画像処理
Pillow>=10.0.0

# HTTP通信
requests>=2.31.0
